move test into correct place
